Include mp support for tiling and tile datasets with non uniform struct
Two major changes:
added support for multiprocessing in tiling. Convert all the functions to collect arguments for tiling which will be replaced by retrieving from a queue in a later PR.
added support for datasets with non-uniform structure .i.e. each samples have varying number of slices, time points etc. Both mask generation and tiling were modified. Mask generation doesn't support mp yet!
