+++ micro_dl/cli/preprocess_script.py
@@ -6,7 +6,8 @@ import time
 
 from micro_dl.input.estimate_flat_field import FlatFieldEstimator2D
 from micro_dl.input.generate_masks import MaskProcessor
-from micro_dl.input.tile_images import ImageTiler
+from micro_dl.input.tile_images_uni_struct import ImageTilerUniform
+from micro_dl.input.tile_images_nonuni_struct import ImageTilerNonUniform
 import micro_dl.utils.aux_utils as aux_utils
 
 
@@ -51,6 +52,10 @@ def pre_process(pp_config):
     time_ids = -1
     if 'time_ids' in pp_config:
         time_ids = pp_config['time_ids']
+    uniform_struct = pp_config['uniform_structure']
+    int2str_len = 3
+    if 'int2str_len' in pp_config:
+        int2str_len = pp_config['int2str_len']
 
     # estimate flat_field images
     correct_flat_field = True if pp_config['correct_flat_field'] else False
@@ -75,6 +80,8 @@ def pre_process(pp_config):
             flat_field_dir=flat_field_dir,
             time_ids=time_ids,
             slice_ids=slice_ids,
+            int2str_len=int2str_len,
+            uniform_struct=uniform_struct
         )
         str_elem_radius = 5
         if 'str_elem_radius' in pp_config['masks']:
@@ -84,22 +91,40 @@ def pre_process(pp_config):
             correct_flat_field=correct_flat_field,
             str_elem_radius=str_elem_radius,
         )
+
         mask_dir = mask_processor_inst.get_mask_dir()
         mask_channel = mask_processor_inst.get_mask_channel()
 
     # Tile frames
     tile_dir = None
     if pp_config['do_tiling']:
+        channel_ids = -1
+        if 'channels' in pp_config['tile']:
+            channel_ids = pp_config['tile']['channels']
+        num_workers = 4
+        if 'num_workers' in pp_config['tile']:
+            num_workers = pp_config['num_workers']
         start = time.time()
-        tile_inst = ImageTiler(
-            input_dir=input_dir,
-            output_dir=output_dir,
-            tile_dict=pp_config['tile'],
-            time_ids=time_ids,
-            slice_ids=slice_ids,
-            flat_field_dir=flat_field_dir,
-            uniform_structure=False
-        )
+        if uniform_struct:
+            tile_inst = ImageTilerUniform(input_dir=input_dir,
+                                          output_dir=output_dir,
+                                          tile_dict=pp_config['tile'],
+                                          time_ids=time_ids,
+                                          slice_ids=slice_ids,
+                                          channel_ids=channel_ids,
+                                          flat_field_dir=flat_field_dir,
+                                          num_workers=num_workers,
+                                          int2str_len=int2str_len)
+        else:
+            tile_inst = ImageTilerNonUniform(input_dir=input_dir,
+                                             output_dir=output_dir,
+                                             tile_dict=pp_config['tile'],
+                                             time_ids=time_ids,
+                                             slice_ids=slice_ids,
+                                             channel_ids=channel_ids,
+                                             flat_field_dir=flat_field_dir,
+                                             num_workers=num_workers,
+                                             int2str_len=int2str_len)
         tile_dir = tile_inst.get_tile_dir()
         # If you're using min fraction, it assumes you've generated masks
         # and want to tile only the ones with a minimum amount of foreground
@@ -110,7 +135,6 @@ def pre_process(pp_config):
                 mask_channel=mask_channel,
             )
         else:
-            start = time.time()
             tile_inst.tile_stack()
         print('Tiling time: {}'.format(time.time() - start))
