+++ micro_dl/utils/image_utils.py
@@ -10,9 +10,6 @@ from skimage.filters import threshold_otsu
 from skimage.morphology import disk, ball, binary_opening
 from skimage.transform import resize
 
-import micro_dl.utils.aux_utils as aux_utils
-import micro_dl.utils.normalize as normalize
-
 
 def resize_image(input_image, output_shape):
     """Resize image to a specified shape
@@ -130,306 +127,10 @@ def fit_polynomial_surface_2D(sample_coords,
     return poly_surface
 
 
-def preprocess_imstack(frames_metadata,
-                       input_dir,
-                       depth,
-                       time_idx,
-                       channel_idx,
-                       slice_idx,
-                       pos_idx,
-                       flat_field_im=None,
-                       hist_clip_limits=None):
-    """
-    Preprocess image given by indices: flatfield correction, histogram
-    clipping and z-score normalization is performed.
-
-    :param int time_idx: Time index
-    :param int channel_idx: Channel index
-    :param int slice_idx: Slice (z) index
-    :param int pos_idx: Position (FOV) index
-    :param np.array flat_field_im: Flat field image for channel
-    :param list hist_clip_limits: Limits for histogram clipping (size 2)
-    :return np.array im: 2D preprocessed image
-    """
-    margin = 0 if depth == 1 else depth // 2
-    im_stack = []
-    for z in range(slice_idx - margin, slice_idx + margin + 1):
-        meta_idx = aux_utils.get_meta_idx(
-            frames_metadata,
-            time_idx,
-            channel_idx,
-            z,
-            pos_idx,
-        )
-        file_path = os.path.join(
-            input_dir,
-            frames_metadata.loc[meta_idx, "file_name"],
-        )
-        im = read_image(file_path)
-        if flat_field_im is not None:
-            im = apply_flat_field_correction(
-                im,
-                flat_field_image=flat_field_im,
-            )
-        im_stack.append(im)
-    # Stack images
-    im_stack = np.stack(im_stack, axis=2)
-    # normalize
-    if hist_clip_limits is not None:
-        im_stack = normalize.hist_clipping(
-            im_stack,
-            hist_clip_limits[0],
-            hist_clip_limits[1],
-        )
-    return normalize.zscore(im_stack)
-
-
-def tile_image(input_image,
-               tile_size,
-               step_size,
-               isotropic=False,
-               return_index=False,
-               min_fraction=None):
-    """
-     Tiles the image based on given tile and step size.
-
-    :param np.array input_image: input image to be tiled
-    :param list/tuple/np array tile_size: size of the blocks to be tiled
-     from the image
-    :param list/tuple/np array step_size: size of the window shift. In case of
-     no overlap, the step size is tile_size. If overlap, step_size < tile_size
-    :param bool isotropic: if 3D, make the grid/shape isotropic
-    :param bool return_index: indicator for returning tile indices
-    :param float min_fraction: Minimum fraction of foreground in mask for
-    including tile
-    :return: a list with tuples of tiled image id of the format
-     rrmin-rmax_ccmin-cmax_slslmin-slmax and tiled image
-     if return_index=True: return a list with tuples of crop indices
-    """
-
-    def check_in_range(cur_value, max_value, tile_size):
-        """Get the start index for edge tiles
-
-        :param int cur_value: cur index in row / col / slice
-        :param int max_value: n_rows, n_cols or n_slices
-        :param int tile_size: size of tile along one dimension (row, col,
-         slice)
-        :return: int start_value - adjusted start_index to fit the edge tile
-        """
-        cur_length = max_value - cur_value
-        miss_length = tile_size - cur_length
-        start_value = cur_value - miss_length
-        return start_value
-
-    def use_tile(cropped_img, min_fraction):
-        """
-        Determine if tile should be used given minimum image foreground fraction
-        :param np.array cropped_img: Image tile
-        :param float min_fraction: Minimum fraction of image being foreground
-        :return bool use_tile: Indicator if tile should be used
-        """
-        use_tile = True
-        if min_fraction is not None:
-            mask_fraction = np.mean(cropped_img)
-            if mask_fraction < min_fraction:
-                use_tile = False
-        return use_tile
-
-    # Add to tile size and step size in case of 3D images
-    im_shape = input_image.shape
-    if len(im_shape) == 3:
-        if len(tile_size) == 2:
-            tile_size.append(im_shape[2])
-        # Step size in z is assumed to be the same as depth
-        if len(step_size) == 2:
-            step_size.append(im_shape[2])
-
-    assert len(tile_size) == len(step_size),\
-        "Tile {} and step size {} mismatch".format(tile_size, step_size)
-    assert np.all(step_size <= tile_size),\
-        "Step size {} > tile size {}".format(step_size, tile_size)
-    assert np.all(tile_size) > 0,\
-        "Tile size must be > 0, not {}".format(tile_size)
-
-    n_rows = input_image.shape[0]
-    n_cols = input_image.shape[1]
-    n_dim = len(input_image.shape)
-    if n_dim == 3:
-        n_slices = input_image.shape[2]
-
-    if isotropic:
-        isotropic_shape = [tile_size[0], ] * len(tile_size)
-        isotropic_cond = not(list(tile_size) == isotropic_shape)
-    else:
-        isotropic_cond = isotropic
-
-    cropped_image_list = []
-    cropping_index = []
-    for row in range(0, n_rows - tile_size[0] + step_size[0], step_size[0]):
-        if row + tile_size[0] > n_rows:
-            row = check_in_range(row, n_rows, tile_size[0])
-        for col in range(0, n_cols - tile_size[1] + step_size[1], step_size[1]):
-            if col + tile_size[1] > n_cols:
-                col = check_in_range(col, n_cols, tile_size[1])
-            img_id = 'r{}-{}_c{}-{}'.format(row, row + tile_size[0],
-                                            col, col + tile_size[1])
-            if n_dim == 3:
-                for sl in range(0, n_slices, step_size[2]):
-                    if sl + step_size[2] > n_slices:
-                        sl = check_in_range(sl, n_slices, tile_size[2])
-
-                    cur_index = (row, row + tile_size[0],
-                                 col, col + tile_size[1],
-                                 sl, sl + tile_size[2])
-                    img_id = '{}_sl{}-{}'.format(img_id, sl, sl + tile_size[2])
-                    cropped_img = input_image[row: row + tile_size[0],
-                                              col: col + tile_size[1],
-                                              sl: sl + tile_size[2]]
-                    if isotropic_cond:
-                        cropped_img = resize_image(cropped_img,
-                                                   isotropic_shape)
-                    if use_tile(cropped_img, min_fraction):
-                        cropped_image_list.append([img_id, cropped_img])
-                        cropping_index.append(cur_index)
-            else:
-                cur_index = (row, row + tile_size[0], col, col + tile_size[1])
-                cropped_img = input_image[row: row + tile_size[0],
-                                          col: col + tile_size[1]]
-                if use_tile(cropped_img, min_fraction):
-                    cropped_image_list.append([img_id, cropped_img])
-                    cropping_index.append(cur_index)
-    if return_index:
-        return cropped_image_list, cropping_index
-    return cropped_image_list
-
-
-def crop_at_indices(input_image, crop_indices, isotropic=False):
-    """Crop image into tiles at given indices
-
-    :param np.array input_image: input image for cropping
-    :param list crop_indices: list of indices for cropping
-    :param bool isotropic: if 3D, make the grid/shape isotropic
-    :return: a list with tuples of cropped image id of the format
-     rrmin-rmax_ccmin-cmax_slslmin-slmax and cropped image
-    """
-
-    n_dim = len(input_image.shape)
-    tiles_list = []
-    im_depth = input_image.shape[2]
-    for cur_idx in crop_indices:
-        img_id = 'r{}-{}_c{}-{}'.format(cur_idx[0], cur_idx[1],
-                                        cur_idx[2], cur_idx[3])
-
-        cropped_img = input_image[cur_idx[0]: cur_idx[1],
-                                  cur_idx[2]: cur_idx[3], ...]
-        if n_dim == 3:
-            img_id = '{}_sl{}-{}'.format(img_id, 0, im_depth)
-
-            if isotropic:
-                img_shape = cropped_img.shape
-                isotropic_shape = [img_shape[0], ] * len(img_shape)
-                cropped_img = resize_image(cropped_img, isotropic_shape)
-
-        tiles_list.append([img_id, cropped_img])
-    return tiles_list
-
-
-def write_tile(file_name_list, flip):
-    """
-    Write tile function that can be called using threading.
-    :param list file_name_list: [file path, tile]
-    """
-    [file_name, tile] = file_name_list
-    if flip:
-        tile = np.transpose(tile, (2, 0, 1))
-    np.save(file_name, tile,  allow_pickle=True, fix_imports=True)
-
-
-def crop_at_indices_save(input_fnames,
-                         flat_field_fname,
-                         hist_clip_limits,
-                         time_idx,
-                         channel_idx,
-                         pos_idx,
-                         slice_idx,
-                         crop_indices,
-                         data_format,
-                         isotropic,
-                         save_dir):
-    """Crop image into tiles at given indices and save
-
-    :param tuple input_fnames: tuple of input fnames with full path
-    :param str flat_field_fname: fname of flat field image
-    :param tuple hist_clip_limits: limits for histogram clipping
-    :param int time_idx: time point of input image
-    :param int channel_idx: channel idx of input image
-    :param int slice_idx: slice idx of input image
-    :param int pos_idx: sample idx of input image
-    :param tuple crop_indices: tuple of indices for cropping
-    :param str data_format: channels_first / channels_last
-    :param bool isotropic: if 3D, make the grid/shape isotropic
-    :param str save_dir: output dir to save tiles
-    :return: a list of dicts with metadata
-    """
-
-    im_stack = []
-    for fname in input_fnames:
-        im = read_image(fname)
-        if flat_field_fname is not None:
-            flat_field_image = np.load(flat_field_fname)
-            im = apply_flat_field_correction(im,
-                                             flat_field_image)
-        im_stack.append(im)
-    input_image = np.stack(im_stack, axis=2)
-    if hist_clip_limits is not None:
-        input_image = normalize.hist_clipping(
-            input_image,
-            hist_clip_limits[0],
-            hist_clip_limits[1]
-        )
-    input_image = normalize.zscore(input_image)
-
-    n_dim = len(input_image.shape)
-    im_depth = input_image.shape[2]
-    tiled_metadata = []
-    for cur_idx in crop_indices:
-        img_id = 'r{}-{}_c{}-{}'.format(cur_idx[0], cur_idx[1],
-                                        cur_idx[2], cur_idx[3])
-
-        cropped_img = input_image[cur_idx[0]: cur_idx[1],
-                                  cur_idx[2]: cur_idx[3], ...]
-        if n_dim == 3 and len(cur_idx) == 6:
-            img_id = '{}_sl{}-{}'.format(img_id, 0, im_depth)
-
-            if isotropic:
-                img_shape = cropped_img.shape
-                isotropic_shape = [img_shape[0], ] * len(img_shape)
-                cropped_img = resize_image(cropped_img, isotropic_shape)
-        file_name = aux_utils.get_im_name(
-            time_idx=time_idx,
-            channel_idx=channel_idx,
-            slice_idx=slice_idx,
-            pos_idx=pos_idx,
-            extra_field=img_id
-            )
-        if data_format == 'channels_first' and len(cropped_img.shape) > 2:
-            cropped_img = np.transpose(cropped_img, (2, 0, 1))
-        np.save(os.path.join(save_dir, file_name),
-                cropped_img,
-                allow_pickle=True,
-                fix_imports=True)
-        tiled_metadata.append({'channel_idx': channel_idx,
-                               'slice_idx': slice_idx,
-                               'time_idx': time_idx,
-                               'file_name': file_name,
-                               'pos_idx': pos_idx,
-                               'row_start': cur_idx[0],
-                               'col_start': cur_idx[2]})
-    return tiled_metadata
-
-
 def create_mask(input_image, str_elem_size=3):
     """Create a binary mask using morphological operations
+
+    Opening removes small objects in the foreground.
 
     :param np.array input_image: generate masks from this image
     :param int str_elem_size: size of the structuring element. typically 3, 5
